% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afths.R
\name{afths}
\alias{afths}
\title{mcmc function}
\usage{
afths(ct, X, nburnin = 1000, nmc = 5000, thin = 1, Xtest = NULL,
  cttest = NULL)
}
\arguments{
\item{ct}{survival response, a \eqn{n*2} matrix with first column as response and second column as right censored indicator,
1 is event time and 0 is right censored.}

\item{X}{Matrix of covariates, dimension \eqn{n*p}.}

\item{nburnin}{number of burin samples to discard}

\item{nmc}{number of markov chain monte carlo samples after burnin}

\item{thin}{thinning parameter. Default is 1 (no thinning)}

\item{Xtest}{test design matrix}

\item{cttest}{test survival response}
}
\value{
\item{BetaHat}{Posterior mean of \eqn{\beta} for survival model, a \eqn{p} by 1 vector}
\item{BetaMedian}{Posterior median of \eqn{beta} for survival model, a \eqn{p} by 1 vector}
\item{Sigma2Hat}{Posterior mean of error variance \eqn{\sigma^2}}
\item{AlphaHat}{Posterior mean of \eqn{\alpha}}
\item{BetaSamples}{Posterior samples of \eqn{\beta}}
\item{Sigma2Samples}{Posterior samples of \eqn{\sigma^2}}
\item{SurvivalHat}{Predictive survival probability}
\item{LogTimeHat}{Predictive log time}
\item{DIC}{Deviance Information Criterion}
\item{WAIC}{Widely Accepted Information Criterion}
\item{LambdaHat}{Posterior mean of \eqn{\lambda}, \eqn{p/2} by 1 vector}
}
\description{
MCMC routine for the strucatural equation model
}
